+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 1.1
 Not yet released.
 
 - Move i18n into core. Deprecate `wtforms.ext.i18n`.
+- Fix issue rendering SelectFields with value=True
 
 Version 1.0.5
 -------------