+++ CHANGES.rst
@@ -1,10 +1,12 @@
 WTForms Changelog
 =================
 
-Version 1.1
+Version 2.0
 -----------
 Not yet released.
 
+- Add new `class Meta` paradigm for much more powerful customization 
+  of WTForms.
 - Move i18n into core. Deprecate `wtforms.ext.i18n`.
 - Fix issue rendering SelectFields with value=True
 - Make `DecimalField` able to use babel locale-based number formatting.